requirements.txt
# Discord Bot Framework
discord.py>=2.0.0

# Async HTTP Client
aiohttp>=3.8.0

# Fuzzy String Matching
rapidfuzz>=3.0.0

# Environment Variable Management
python-dotenv>=1.0.0
